Update B2MML-equipment.xsd
EquipmentInformationType: EquipmentCapabilityTestSpecification (line 165) should be deprecated and replaced by TestSpecification to align with updated Part 2, Operations Test Model.  This also aligns with the OpEquipmentRequirementType: TestSpecificationID in the Common Schema for use in OperationsDefinition.

EquipmentType:  EquipmentCapabilityTestSpecificationID (line 193) should be deprecated and replaced by TestSpecificationID to align with updated Part 2, Operations Test Model.  This also aligns with the OpEquipmentRequirementType: TestSpecificationID in the Common Schema for use in OperationsDefinition.

EquipmentPropertyType: Delete TestResult (line 211). In updated Part 2, Test Result was removed from the equipment property object and included in the Operations Test Model.  Test Result object has an association relationship to Equipment Actual and shall be added to the OpEquipmentActual schema in the Common Schema.  Also, Test Result object is now a 1st order object and should be added as a common schema with transactions.

EquipmentClassType:  EquipmentCapabilityTestSpecificationID (line 236) should be deprecated and replaced by TestSpecificationID to align with updated Part 2, Operations Test Model.  This also aligns with the OpEquipmentRequirementType: TestSpecificationID in the Common Schema for use in OperationsDefinition.