Create B2MML-OperationsTest.xsd
This new schema creates an OperationsTest.xsd schema that has test specification type and test result type and allow 1st order transactions so that it aligns with updated Part 2.  This schema replaces TestSpecification.xsd and TestResultType in Common Schema.  I do not believe that the 1st order transactions were added to the updated part 5. But the updated part 2 as part of the collaboration with OAGi viewed Test Results as a 1st order object and transactions.
B2MML 7.0 beta has created a B2MML-TestSpecification.xsd schema and created the complex types of "TestResultType" and TestPropertyMeasurementType in the Common Schema.
B2MML 7.0 still has the TestResult type as an attribute in each instance resource property per the old version of ISA-95 resource models and B2MML 6.0. This was a composite relationship which did not allow Test Result to be exchanged and queried as a 1st class object.
In the new Operations Test Model in updated Part 2, Test Result is a 1st order object that is able to be exchanged/queried.
The MESA Committee may reject changing their implementation. If so,