Update B2MML-PhysicalAsset.xsd
ERDI, MESA:  PhysicalAssetPropertyType: 
DELETE TestResult . In updated Part 2, Test Result was removed from the equipment property object and included in the Operations Test Model.  Test Result object has an association relationship to Equipment Actual and shall be added to the OpEquipmentActual schema in the Common Schema.  Also, Test Result object is now a 1st order object and should be added as a common schema with transactions.
#7