Update B2MML-OperationalLocation.xsd
ERDI, MESA: #2 
**Background**
In updated 950002 and 950004 models,  95 Committee agreed that the standard's convention is show the logical direction of all UML model relationships. Most relationships in models are unidirectional from a source object to a target object where the UML and XML best practice is applied with only the naming of target role element in the relationship role table for each object.  
The current B2MML 7.0 beta is very inconsistent by having both source and target elements for unidirectional relationships in many types and having no source element in many types.
**Solution**
In B2MML: Change the name of existing source element of unidirectional relationship in schema to a naming convention of XXXSourceID from XXXXID. 

CHANGE:       "OperationalLocationClassType"  
<xsd:element name = "OperationalLocationID"       type = "IdentifierType"  minOccurs = "0" maxOccurs = "unbounded"/>
TO: "OperationalLocationSourceID"